// lib/server/plugin/types.ts
export interface ServerConfig {
	modType: string;
	serverDir: string;
	minecraftVersion: string;
	loaderType: string;
	pluginDir: string;
	tag: string | null;
	port: number[];
	apiPort: number | null;
}